Revert "DTE-317: latest tdr changes to bagit fields handled in bagit to dri sip"
